Replace Hugo with Mkdocs.
