--- .github/actions/setup-hugo/action.yml
@@ -1,30 +1,0 @@
-name: "Setup Hugo"
-description: "Setup Hugo, Pandoc, and Node for Docsy"
-runs:
-  using: composite
-  steps:
-    - name: Setup Hugo (Linux)
-      uses: peaceiris/actions-hugo@v2
-      with:
-        hugo-version: '0.112.7'
-        extended: true
-      if: runner.os == 'Linux'
-    - name: Setup Hugo (macOS)
-      run: brew install hugo
-      shell: bash
-      if: runner.os == 'macOS'
-    - name: Install Scoop (Windows)
-      if: runner.os == 'Windows'
-      uses: MinoruSekine/setup-scoop@main
-    - name: Setup Hugo (Windows)
-      run: scoop install hugo-extended
-      shell: bash
-      if: runner.os == 'Windows'
-    - name: Setup Node
-      uses: actions/setup-node@v4
-      with:
-        node-version: '16'
-    - name: Setup Pandoc
-      uses: r-lib/actions/setup-pandoc@v2
-      with:
-        pandoc-version: '2.19.2'
