+++ .github/workflows/build-and-test.yaml
@@ -24,8 +24,6 @@ jobs:
         uses: bigladder/github-actions/setup-python-poetry@main
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Setup Hugo
-        uses: ./.github/actions/setup-hugo
       - name: Build and test
         run: poetry run doit
       # - name: Upload Build Content
