+++ .github/workflows/release.yaml
@@ -24,8 +24,6 @@ jobs:
         uses: bigladder/github-actions/setup-python-poetry@main
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Setup Hugo
-        uses: ./.github/actions/setup-hugo
       - name: Build and publish package
         run: |
           poetry config pypi-token.pypi ${{ secrets.PYPI_TOKEN }}
