+++ dodo.py
@@ -1,5 +1,5 @@
 from lattice import Lattice
-import os
+from pathlib import Path
 
 from doit.tools import create_folder
 
@@ -9,26 +9,27 @@ BUILD_PATH = "build"
 
 examples = []
 
-for example_dir in sorted(os.listdir(EXAMPLES_PATH)):
-  example_dir_path = os.path.join(EXAMPLES_PATH, example_dir)
-  if os.path.isdir(example_dir_path):
-    build_dir_path = os.path.join(BUILD_PATH, example_dir)
+for example_dir in sorted([str(dir_name) for dir_name in Path(EXAMPLES_PATH).iterdir()]):
+  example_dir_path = Path(example_dir)
+  example_name = example_dir_path.name
+  if example_dir_path.is_dir():
+    build_dir_path = Path(BUILD_PATH, example_name)
     create_folder(build_dir_path)
     examples.append(Lattice(example_dir_path, build_dir_path, build_output_directory_name=None, build_validation=False))
 
-BASE_META_SCHEMA_PATH = os.path.join(SOURCE_PATH, "meta.schema.yaml")
-CORE_SCHEMA_PATH = os.path.join(SOURCE_PATH, "core.schema.yaml")
+BASE_META_SCHEMA_PATH = Path(SOURCE_PATH, "meta.schema.yaml")
+CORE_SCHEMA_PATH = Path(SOURCE_PATH, "core.schema.yaml")
 
 def task_generate_meta_schemas():
   '''Generate JSON meta schemas'''
   for example in examples:
-    name = os.path.basename(example.root_directory)
+    name = Path(example.root_directory).name
     yield {
       'name': name,
       'file_dep': [schema.path for schema in example.schemas] +
                   [BASE_META_SCHEMA_PATH,
                    CORE_SCHEMA_PATH,
-                   os.path.join(SOURCE_PATH, "meta_schema.py")],
+                   Path(SOURCE_PATH, "meta_schema.py")],
       'targets': [schema.meta_schema_path for schema in example.schemas],
       'actions': [
         (example.generate_meta_schemas, [])
@@ -40,7 +41,7 @@ def task_generate_meta_schemas():
 def task_validate_schemas():
   '''Validate the example schemas against the JSON meta schema'''
   for example in examples:
-    name = os.path.basename(example.root_directory)
+    name = Path(example.root_directory).name
     yield {
       'name': name,
       'task_dep': [f"generate_meta_schemas:{name}"],
@@ -48,7 +49,7 @@ def task_validate_schemas():
                   [schema.meta_schema_path for schema in example.schemas] +
                   [BASE_META_SCHEMA_PATH,
                    CORE_SCHEMA_PATH,
-                   os.path.join(SOURCE_PATH, "meta_schema.py")],
+                   Path(SOURCE_PATH, "meta_schema.py")],
       'actions': [
         (example.validate_schemas, [])
       ]
@@ -57,7 +58,7 @@ def task_validate_schemas():
 def task_generate_json_schemas():
   '''Generate JSON schemas'''
   for example in examples:
-    name = os.path.basename(example.root_directory)
+    name = Path(example.root_directory).name
     yield {
       'name': name,
       'task_dep': [f"validate_schemas:{name}"],
@@ -65,7 +66,7 @@ def task_generate_json_schemas():
                   [schema.meta_schema_path for schema in example.schemas] +
                   [CORE_SCHEMA_PATH,
                    BASE_META_SCHEMA_PATH,
-                   os.path.join(SOURCE_PATH, "schema_to_json.py")],
+                   Path(SOURCE_PATH, "schema_to_json.py")],
       'targets': [schema.json_schema_path for schema in example.schemas],
       'actions': [
         (example.generate_json_schemas, [])
@@ -76,10 +77,10 @@ def task_generate_json_schemas():
 def task_validate_example_files():
   '''Validate example files against JSON schema'''
   for example in examples:
-    name = os.path.basename(example.root_directory)
+    name = Path(example.root_directory).name
     yield {
       'name': name,
-      'file_dep': [schema.json_schema_path for schema in example.schemas] + example.examples + [os.path.join(SOURCE_PATH, "schema_to_json.py")],
+      'file_dep': [schema.json_schema_path for schema in example.schemas] + example.examples + [Path(SOURCE_PATH, "schema_to_json.py")],
       'task_dep': [f"generate_json_schemas:{name}"],
       'actions': [
         (example.validate_example_files, [])
@@ -89,11 +90,11 @@ def task_validate_example_files():
 def task_generate_markdown():
   '''Generate markdown documentation from templates'''
   for example in examples:
-    name = os.path.basename(example.root_directory)
+    name = Path(example.root_directory).name
     yield {
       'name': name,
       'targets': [template.markdown_output_path for template in example.doc_templates],
-      'file_dep': [schema.path for schema in example.schemas] + [template.path for template in example.doc_templates] + [os.path.join(SOURCE_PATH, "docs", "grid_table.py")],
+      'file_dep': [schema.path for schema in example.schemas] + [template.path for template in example.doc_templates] + [Path(SOURCE_PATH, "docs", "grid_table.py")],
       'task_dep': [f"validate_schemas:{name}"],
       'actions': [
         (example.generate_markdown_documents, [])
@@ -104,7 +105,7 @@ def task_generate_markdown():
 def task_generate_cpp_code():
     '''Generate CPP headers and source for example schema.'''
     for example in examples:
-        name = os.path.basename(example.root_directory)
+        name = Path(example.root_directory).name
         yield {
           'name': name,
           'task_dep': [f"validate_schemas:{name}"],
@@ -112,7 +113,7 @@ def task_generate_cpp_code():
                       [schema.meta_schema_path for schema in example.schemas] +
                       [CORE_SCHEMA_PATH,
                       BASE_META_SCHEMA_PATH,
-                      os.path.join(SOURCE_PATH, "header_entries.py")],
+                      Path(SOURCE_PATH, "header_entries.py")],
           'targets': [schema.cpp_header_path for schema in example.cpp_schemas] +
                      [schema.cpp_source_path for schema in example.cpp_schemas] +
                      example.cpp_support_headers(),
@@ -125,12 +126,12 @@ def task_generate_cpp_code():
 def task_generate_web_docs():
   '''Generate markdown documentation from templates'''
   for example in examples:
-    name = os.path.basename(example.root_directory)
+    name = Path(example.root_directory).name
     yield {
       'name': name,
       'task_dep': [f"validate_schemas:{name}",f"generate_json_schemas:{name}",f"validate_example_files:{name}"],
-      'file_dep': [schema.path for schema in example.schemas] + [template.path for template in example.doc_templates],
-      'targets': [os.path.join(example.web_docs_directory_path,"public")],
+      'file_dep': [schema.path for schema in example.schemas] + [template.path for template in example.doc_templates] + [Path(SOURCE_PATH, "docs", "mkdocs_web.py")],
+      'targets': [Path(example.web_docs_directory_path,"public")],
       'actions': [
         (example.generate_web_documentation, [])
       ],
