+++ examples/fan_spec/docs/web/about.md
@@ -1,17 +1,9 @@
-{{% blocks/cover title="About Fan Data Model" height="auto" %}}
 This is an example of how to put together a Lattice data model.
-{{% /blocks/cover %}}
 
-{{% blocks/section %}}
 ## Authors
 
 Lattice is developed by [Big Ladder Software](https://bigladdersoftware.com/).
 
-{{% /blocks/section %}}
-
-{{% blocks/section  %}}
 ## Acknowledgements
 
 This is developed with support from IBPSA-USA.
-
-{{% /blocks/section %}}
