+++ examples/fan_spec/docs/web/favicon.ico
