--- examples/fan_spec/docs/web/landing.md
@@ -1,1 +1,0 @@
-An open source data model specification for fans (based on ASHRAE 205).
