--- lattice/docs/hugo_layouts/_default/single.html
@@ -1,9 +1,0 @@
-{{ define "main" }}
-{{ with .Content }}
-<main id="main">
-  <div class="td-content">
-    {{ . }}
-  </div>
-</main>
-{{ end }}
-{{ end }}
