--- lattice/docs/hugo_layouts/specifications/list.html
@@ -1,21 +1,0 @@
-{{ define "main" }}
-<div class="td-content">
-	<header class="article-meta">
-		{{ partial "taxonomy_terms_article_wrapper.html" . }}
-		{{ if (and (not .Params.hide_readingtime) (.Site.Params.ui.readingtime.enable)) }}
-			{{ partial "reading-time.html" . }}
-		{{ end }}
-	</header>
-	{{ .Content }}
-        {{ partial "section-index.html" . }}
-	{{ if (and (not .Params.hide_feedback) (.Site.Params.ui.feedback.enable) (.Site.GoogleAnalytics)) }}
-		{{ partial "feedback.html" .Site.Params.ui.feedback }}
-		<br />
-	{{ end }}
-	{{ if (.Site.DisqusShortname) }}
-		<br />
-		{{ partial "disqus-comment.html" . }}
-	{{ end }}
-	{{ partial "page-meta-lastmod.html" . }}
-</div>
-{{ end }}
