--- lattice/docs/hugo_layouts/specifications/list.print.html
@@ -1,3 +1,0 @@
-{{ define "main" }}
-{{ partial "print/render"  . }}
-{{ end }}
