--- lattice/docs/hugo_layouts/specifications/single.html
@@ -1,21 +1,0 @@
-{{ define "main" }}
-<div class="td-content">
-	{{ with .Params.description }}<div class="lead">{{ . | markdownify }}</div>{{ end }}
-	<header class="article-meta">
-		{{ partial "taxonomy_terms_article_wrapper.html" . }}
-		{{ if (and (not .Params.hide_readingtime) (.Site.Params.ui.readingtime.enable)) }}
-			{{ partial "reading-time.html" . }}
-		{{ end }}
-	</header>
-	{{ .Content }}
-	{{ if (and (not .Params.hide_feedback) (.Site.Params.ui.feedback.enable) (.Site.GoogleAnalytics)) }}
-		{{ partial "feedback.html" .Site.Params.ui.feedback }}
-		<br />
-	{{ end }}
-	{{ if (.Site.Params.DisqusShortname) }}
-		<br />
-		{{ partial "disqus-comment.html" . }}
-	{{ end }}
-	{{ partial "page-meta-lastmod.html" . }}
-</div>
-{{ end }}
