--- lattice/docs/scss_template.scss.j2
@@ -1,21 +1,0 @@
-$primary: {{primary}}  !default;
-$secondary: {{secondary}} !default;
-
-
-
-$navbar-dark-color:                 rgba({{menu_text}}, .5) !default;
-$navbar-dark-hover-color:           rgba({{menu_text}}, .75) !default;
-$navbar-dark-active-color:          {{menu_text}} !default;
-$navbar-dark-disabled-color:        rgba({{menu_text}}, .25) !default;
-$navbar-dark-toggler-border-color:  rgba({{menu_text}}, .1) !default;
-
-$navbar-light-color:                rgba({{menu_text}}, .5) !default;
-$navbar-light-hover-color:          rgba({{menu_text}}, .7) !default;
-$navbar-light-active-color:         rgba({{menu_text}}, .9) !default;
-$navbar-light-disabled-color:       rgba({{menu_text}}, .3) !default;
-$navbar-light-toggler-border-color: rgba({{menu_text}}, .1) !default;
-
-$navbar-light-brand-color:                $navbar-light-active-color !default;
-$navbar-light-brand-hover-color:          $navbar-light-active-color !default;
-$navbar-dark-brand-color:                 $navbar-dark-active-color !default;
-$navbar-dark-brand-hover-color:           $navbar-dark-active-color !default;
