+++ pyproject.toml
@@ -13,15 +13,17 @@ python = "^3.8"
 jsonschema = "*"
 pyyaml = "*"
 cbor2 = "*"
-Jinja2 = "*"
+Jinja2 = "^3.1.4"
 stringcase = "1.2.0"
 pygit2 = "^1.13.0"
+mkdocs-material = "*"
+markdown-grid-tables = "*"
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.dev.dependencies]
 pytest = "^7.1.3"
 doit = "*"
 pylint = "*"
-tomlkit = "0.11.8"
+tomli = "2.0.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
